chore: Introducing FirebaseAdmin.Auth.Jwt namespace for internal APIs
