+++ FirebaseAdmin/FirebaseAdmin.Tests/Auth/ImportUserRecordArgsTest.cs
@@ -16,6 +16,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using FirebaseAdmin.Auth.Jwt;
 using Google.Apis.Json;
 using Newtonsoft.Json;
 using Xunit;
