+++ FirebaseAdmin/FirebaseAdmin/Auth/Jwt/FirebaseTokenFactory.cs
@@ -27,7 +27,7 @@ using Google.Apis.Util;
 "3003684e85e61cf15f13150008c81f0b75a252673028e530ea95d0c581378da8c6846526ab9597" +
 "4c6d0bc66d2462b51af69968a0e25114bde8811e0d6ee1dc22d4a59eee6a8bba4712cba839652f" +
 "badddb9c")]
-namespace FirebaseAdmin.Auth
+namespace FirebaseAdmin.Auth.Jwt
 {
     /// <summary>
     /// A helper class that creates Firebase custom tokens.
