+++ FirebaseAdmin/FirebaseAdmin/Auth/Jwt/FixedAccountIAMSigner.cs
@@ -19,7 +19,7 @@ using Google.Apis.Auth.OAuth2;
 using Google.Apis.Http;
 using Google.Apis.Util;
 
-namespace FirebaseAdmin.Auth
+namespace FirebaseAdmin.Auth.Jwt
 {
     /// <summary>
     /// An <see cref="ISigner"/> implementation that uses the IAM service to sign data. Unlike
