+++ FirebaseAdmin/FirebaseAdmin/Auth/Jwt/ServiceAccountSigner.cs
@@ -18,7 +18,7 @@ using System.Threading.Tasks;
 using Google.Apis.Auth.OAuth2;
 using Google.Apis.Util;
 
-namespace FirebaseAdmin.Auth
+namespace FirebaseAdmin.Auth.Jwt
 {
     /// <summary>
     /// An <see cref="ISigner"/> implementation that uses service account credentials to sign
