+++ FirebaseAdmin/FirebaseAdmin/Auth/UserRecordArgs.cs
@@ -16,6 +16,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Text.RegularExpressions;
+using FirebaseAdmin.Auth.Jwt;
 using Google.Apis.Json;
 using Newtonsoft.Json;
